+++ src/react/data/QueryData.ts
@@ -72,8 +72,6 @@ export class QueryData<TData, TVariables> extends OperationData<
 
     this.updateObservableQuery();
 
-    if (this.isMounted) this.startQuerySubscription();
-
     return this.getExecuteSsrResult() || this.getExecuteResult();
   }
 
@@ -100,12 +98,21 @@ export class QueryData<TData, TVariables> extends OperationData<
 
   public afterExecute({ lazy = false }: { lazy?: boolean } = {}) {
     this.isMounted = true;
+    const options = this.getOptions();
+    const ssrDisabled = options.ssr === false;
+    if (
+      this.currentObservable &&
+      !ssrDisabled &&
+      !this.ssrInitiated()
+    ) {
+      this.startQuerySubscription();
+    }
 
     if (!lazy || this.runLazy) {
       this.handleErrorOrCompleted();
     }
 
-    this.previousOptions = this.getOptions();
+    this.previousOptions = options;
     return this.unmount.bind(this);
   }
 
@@ -148,12 +155,6 @@ export class QueryData<TData, TVariables> extends OperationData<
     this.onNewData();
   };
 
-  private getExecuteResult(): QueryResult<TData, TVariables> {
-    const result = this.getQueryResult();
-    this.startQuerySubscription();
-    return result;
-  };
-
   private getExecuteSsrResult() {
     const { ssr, skip } = this.getOptions();
     const ssrDisabled = ssr === false;
@@ -177,7 +178,7 @@ export class QueryData<TData, TVariables> extends OperationData<
     }
 
     if (this.ssrInitiated()) {
-      const result = this.getQueryResult() || ssrLoading;
+      const result = this.getExecuteResult() || ssrLoading;
       if (result.loading && !skip) {
         this.context.renderPromises!.addQueryPromise(this, () => null);
       }
@@ -334,7 +335,7 @@ export class QueryData<TData, TVariables> extends OperationData<
     }
   }
 
-  private getQueryResult = (): QueryResult<TData, TVariables> => {
+  private getExecuteResult(): QueryResult<TData, TVariables> {
     let result = this.observableQueryFields() as QueryResult<TData, TVariables>;
     const options = this.getOptions();